update evaluation summary type def, move report typedef file
