+++ app/benchmarking/report/components/_tests_/mocks/mockBenchmarkResultWithMLCCookbook.ts
@@ -1,4 +1,4 @@
-import { CookbooksBenchmarkResult } from '@/app/benchmarking/types/benchmarkReportTypes';
+import { CookbooksBenchmarkResult } from '@/app/benchmarking/report/types/benchmarkReportTypes';
 
 export const mockBenchmarkResultWithMLCCookbook: CookbooksBenchmarkResult = {
   metadata: {
