+++ app/benchmarking/report/components/mlcReportComponents/mlcAISafetyCookbookReportCard.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { RecipeGradeBadge } from '@/app/benchmarking/report/components/badge';
 import { gradeColorsMlc } from '@/app/benchmarking/report/components/gradeColors';
 import { gradingLettersMlcMap } from '@/app/benchmarking/report/components/mlcReportComponents/constants';
-import { CookbookResult } from '@/app/benchmarking/types/benchmarkReportTypes';
+import { CookbookResult } from '@/app/benchmarking/report/types/benchmarkReportTypes';
 import { Icon, IconName } from '@/app/components/IconSVG';
 import { colors } from '@/app/customColors';
 import MlcAiSafetyRecipeRatingResult from './mlcAiSafetyRecipeRatingResult';
