+++ app/benchmarking/utils/calcTotalPromptsByEndpoint.ts
@@ -1,4 +1,4 @@
-import { CookbooksBenchmarkResult } from '@/app/benchmarking/types/benchmarkReportTypes';
+import { CookbooksBenchmarkResult } from '@/app/benchmarking/report/types/benchmarkReportTypes';
 
 export function calcTotalPromptsByEndpoint(
   result: CookbooksBenchmarkResult,
